refactor: modularize conversion tracking interfaces into dedicated files for improved organization and maintainability
